backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
anthropic
openai
celery
redis
boto3
python-dotenv
psycopg2-binary
razorpay
sentry-sdk[fastapi]
supabase
langchain
langchain-community
pymupdf
pytest
httpx
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
